+++ app/schemas/tolerance_value/repository_schemas.py
@@ -1,6 +1,7 @@
 """ToleranceMatch db schemas."""
 
 from datetime import datetime
+from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict
@@ -12,8 +13,8 @@ class ToleranceValueRepoSchema(BaseModel):
     id: UUID
     tolerance_id: UUID
     range_id: UUID
-    upper_value: float
-    lower_value: float
+    upper_value: Decimal
+    lower_value: Decimal
     created_at: datetime
     updated_at: datetime | None
