+++ app/schemas/tolerance_value/response_schemas.py
@@ -1,10 +1,12 @@
 """ToleranceMatch response schemas."""
 
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class ToleranceValueResponseSchema(BaseModel):
     """The common response tolerance value schema."""
 
-    upper_value: float
-    lower_value: float
+    upper_value: Decimal
+    lower_value: Decimal
