+++ app/utils/calculators.py
@@ -1,6 +1,7 @@
 """Tolerance calculators."""
 
 import math
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from app.schemas import ToleranceValueRepoSchema
@@ -11,15 +12,15 @@ if TYPE_CHECKING:
 
 async def calculate_deviations_localized(
     i18n: "TranslatorRunner",
-    target_value: float,
+    target_value: Decimal,
     tolerance_value: type[ToleranceValueRepoSchema],
 ) -> str:
     """Calculate dimensional deviations."""
 
-    def fmt(val: float) -> str:
+    def fmt(val: Decimal) -> str:
         return f"{val:.10f}".rstrip("0").rstrip(".").replace(".", ",")
 
-    def count_decimal_places(value: float) -> int:
+    def count_decimal_places(value: Decimal) -> int:
         s = f"{value:.10f}".rstrip("0")
         if "." in s:
             return len(s.split(".")[1])
@@ -45,29 +46,29 @@ async def calculate_deviations_localized(
 
 
 async def calculate_spindle_speed(
-    cutting_speed: float,
-    tool_diameter: float,
-) -> float:
+    cutting_speed: Decimal,
+    tool_diameter: Decimal,
+) -> Decimal:
     """Calculates the spindle speed (rpm) based on the
     cutting speed and diameter."""
-    return round((1000 * cutting_speed) / (math.pi * tool_diameter), 0)
+    return round((1000 * cutting_speed) / (Decimal(math.pi) * tool_diameter), 0)
 
 
 async def calculate_feed_rate(
-    spindle_speed: float,
+    spindle_speed: Decimal,
     number_of_teeth: int,
-    feed_per_tooth: float,
-) -> float:
+    feed_per_tooth: Decimal,
+) -> Decimal:
     """Calculates the minute feed (mm/min) based on rotation speed,
     number of teeth and feed per tooth."""
     return round((spindle_speed * number_of_teeth * feed_per_tooth), 0)
 
 
 async def calculate_feed_per_tooth(
-    feed_rate: float,
-    spindle_speed: float,
+    feed_rate: Decimal,
+    spindle_speed: Decimal,
     number_of_teeth: int,
-) -> float:
+) -> Decimal:
     """Calculates the feed per tooth (mm/tooth) if the minute feed,
     rotation speed and number of teeth are known."""
     if spindle_speed <= 0 or number_of_teeth <= 0:
@@ -78,9 +79,9 @@ async def calculate_feed_per_tooth(
 
 
 async def calculate_cutting_speed(
-    spindle_speed: float,
-    tool_diameter: float,
-) -> float:
+    spindle_speed: Decimal,
+    tool_diameter: Decimal,
+) -> Decimal:
     """Calculates the cutting speed (m/min) if the rotation speed
     and diameter of the milling cutter are known."""
-    return round((math.pi * tool_diameter * spindle_speed) / 1000, 1)
+    return round((Decimal(math.pi) * tool_diameter * spindle_speed) / 1000, 1)
