+++ app/utils/parsers.py
@@ -1,6 +1,7 @@
 """Parsers module."""
 
 import re
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
@@ -11,7 +12,7 @@ from core.exceptions.base_exceptions import BadRequestError
 
 async def parse_tolerance(
     input_row: str, i18n: "TranslatorRunner"
-) -> tuple[float, str, int]:
+) -> tuple[Decimal, str, int]:
     """Parse tolerance parameters from input row."""
     pattern = r"^(\d+(?:[.,]\d+)?)[ ]*([A-Za-z]+)[ ]*(\d+)$"
     match = re.match(pattern, input_row.strip())
@@ -22,7 +23,7 @@ async def parse_tolerance(
         )
 
     size_str = match.group(1).replace(",", ".")
-    size = float(size_str)
+    size = Decimal(size_str)
     letter = match.group(2)
     quality = int(match.group(3))
